Move further components to common directory
