+++ src/public/common/Intersection.js
