+++ src/public/common/LongText.js
