+++ src/public/common/SectionAvatar.js
