+++ src/public/common/SideMenu.js
