+++ src/public/common/SideMenuItem.js
